fix(gw-panel): 修复审计链的 status_code 解码与 jsonb 哈希往返
audit_chain 的 6 条 --ignored 集成测试原本全挂，暴露两个既存 bug：

1. status_code 解码类型错：operation_logs.status_code 是 bigint，但校验路径的
   AuditRow 按 i32 解码，sqlx 报 mismatched types。按 CONTRACT §3.5 全线改 i64
   （OperationEntry.status_code、AuditRow、oplog 绑参）。canonical 用 to_string()，
   哈希字节不变。

2. jsonb 哈希往返对不上：写路径按 serde 紧凑字节算 HMAC，但列是 jsonb，Postgres
   会重排键、加空白；verify 读的是 metadata::text（归一化后），于是任何带 metadata
   的行都验不过。改成写入前用 Postgres 自己（SELECT $1::jsonb::text）把 metadata
   归一化再做 HMAC —— 与 verify 侧对齐，也与既有实现（同样读 ::text）跨实现一致。

audit_chain 7/7、gw-panel 集成 48/48 全绿。

+++ crates/gw-panel/src/ops/audit_log.rs
@@ -64,7 +64,7 @@ struct AuditRow {
     target: Option<String>,
     method: Option<String>,
     path: Option<String>,
-    status_code: Option<i32>,
+    status_code: Option<i64>,
     ip_address: Option<String>,
     request_id: Option<String>,
     metadata: Option<String>,
